refactor: extract startup runtime hook boundary
